[LibPrologix] Ensure send_line() does not accept commands with new lines
This change was made becuase adding new lines in the command text
is will cause multiple commands to be send at once. This will break
error reporting. In addition, with some instruments (e.g. Agilent
AWG) sending empty lines without any command can cause a desync of
the communication or semi-bogus errors being reported.

Now, the send_line() will try to strip leading & trailing new lines
and spaces. It will also raise an exception when new lines are still
found in the command. This is to stay user-friendly and not raise
exceptions when someone simply added a trailing new line, but still
warn the user if they tried to do something unsupported.