[LibAWG] Add error handling & safe call methods
This commit introduces a higher-level interface for sending
random commands to the AWG. Newly introduced methods with _awg()
suffix will handle errors by default.

The new interface will not only send the desired commend to the
instrument, but will also always query for remotely-generated
errors and raise them as needed. This method is prefered over using
"raw" non-_awg() suffixed methods as it ensures commands and queries
send to the AWG are valid and there are no surprised. This is
especially important when configuring more complex behaviors like
arbitrary waveforms, where one wrong command can have an effect on
other subsequent commands.

In addition, since erors are storead as a FIFO on the instrument,
the queue will be read on startup and reported. This will allow for
easier remote operation without operating the instrument's front
panel.