[LibAWG] Ensure predictible AWG state upon connection
Connecting via GPIB to the AWG does not reset its state
in any way. In other words, whatever was configured using the
front panel of the instrument or via GPIB session before will
be in effect. This can cause issues with our code expecting
clean state to be present (e.g. output mode). This commit
ensures that the AWG is reset (which takes ~1 second) to its
clean state.

In addition, the state is communicated via the instrument's
front panel display to avoid confustion.